Поток для сборки мусора с пониженным приоритетом

+++ time.h
@@ -1,7 +1,8 @@
 #ifndef SC6_TIME_H
 #define SC6_TIME_H
 
-double getStartTime();
-double getDiffTime(double startTime);
+unsigned long getStartTime();
+
+unsigned long getDiffTime(unsigned long startTime);
 
 #endif //SC6_TIME_H
